add sortlist component
