+++ src/camundaFormIOTasklist.scss
@@ -132,17 +132,6 @@
     width: 13rem;
   }
 
-  .cft-sort-items {
-    border: 1px solid #949494;
-    border-radius: 3%;
-    position: absolute;
-    background-color: #fff;
-    background-clip: padding-box;
-    z-index: 5;
-    padding: 5px;
-    width: 13rem;
-  }
-
   .cft-filter-dropdown-content 
   {
     display: none;
@@ -241,6 +230,8 @@
   .cftf-add-sorting {
     padding-top: 5px;
     margin-top: 5px;
+    position: relative;
+    float: left;
   }
 
   .cftf-arrow {
