+++ src/components/tasklist-sortoptions.vue
@@ -1,6 +1,10 @@
 <template>
     <div v-if="showSortListDropdown" class="cft-sort-items">
-        <div v-for="s in sortOptions" :key="s.sortBy" @click="addSort(s)">
+        <div
+          v-for="s in sortOptions" :key="s.sortBy"
+          @click="addSort(s)"
+          class="mb-2 cft-sort-element"
+        >
             {{s.label}}
         </div>
     </div>
